+++ src/main/java/org/truffleruby/core/basicobject/BasicObjectNodes.java
@@ -10,7 +10,8 @@
 package org.truffleruby.core.basicobject;
 
 import com.oracle.truffle.api.RootCallTarget;
-import com.oracle.truffle.api.dsl.Fallback;
+import com.oracle.truffle.api.interop.InteropLibrary;
+import com.oracle.truffle.api.interop.UnsupportedMessageException;
 import com.oracle.truffle.api.object.Shape;
 import org.truffleruby.Layouts;
 import org.truffleruby.builtins.CoreMethod;
@@ -36,6 +37,7 @@ import org.truffleruby.core.proc.RubyProc;
 import org.truffleruby.core.rope.Rope;
 import org.truffleruby.core.rope.RopeOperations;
 import org.truffleruby.core.symbol.RubySymbol;
+import org.truffleruby.interop.TranslateInteropExceptionNode;
 import org.truffleruby.language.ImmutableRubyObject;
 import org.truffleruby.language.LexicalScope;
 import org.truffleruby.language.Nil;
@@ -149,16 +151,42 @@ public abstract class BasicObjectNodes {
             return Double.doubleToRawLongBits(a) == Double.doubleToRawLongBits(b);
         }
 
-        @Specialization(guards = { "a.getClass() == b.getClass()", "!isPrimitive(a)" }) // since a and b have the same class, implies !isPrimitive(b)
-        protected boolean equalSameClassNonPrimitive(Object a, Object b) {
+        @Specialization(guards = { "isNonPrimitiveRubyObject(a)", "isNonPrimitiveRubyObject(b)" })
+        protected boolean equalRubyObjects(Object a, Object b) {
             return a == b;
         }
 
-        @Fallback
-        protected boolean fallback(Object a, Object b) {
+        @Specialization(guards = { "isNonPrimitiveRubyObject(a)", "isPrimitive(b)" })
+        protected boolean rubyObjectPrimitive(Object a, Object b) {
             return false;
         }
 
+        @Specialization(guards = { "isPrimitive(a)", "isNonPrimitiveRubyObject(b)" })
+        protected boolean primitiveRubyObject(Object a, Object b) {
+            return false;
+        }
+
+        @Specialization(guards = { "isPrimitive(a)", "isPrimitive(b)", "!comparablePrimitives(a, b)" })
+        protected boolean nonComparablePrimitives(Object a, Object b) {
+            return false;
+        }
+
+        @Specialization(guards = "isForeignObject(a) || isForeignObject(b)", limit = "getInteropCacheLimit()")
+        protected boolean equalForeign(Object a, Object b,
+                @CachedLibrary("a") InteropLibrary lhsInterop,
+                @CachedLibrary("b") InteropLibrary rhsInterop) {
+            return lhsInterop.isIdentical(a, b, rhsInterop);
+        }
+
+        protected static boolean isNonPrimitiveRubyObject(Object object) {
+            return object instanceof RubyDynamicObject || object instanceof ImmutableRubyObject;
+        }
+
+        protected static boolean comparablePrimitives(Object a, Object b) {
+            return (a instanceof Boolean && b instanceof Boolean) ||
+                    (RubyGuards.isImplicitLong(a) && RubyGuards.isImplicitLong(b)) ||
+                    (RubyGuards.isImplicitDouble(a) && RubyGuards.isImplicitDouble(b));
+        }
     }
 
     @GenerateUncached
@@ -261,14 +289,19 @@ public abstract class BasicObjectNodes {
             return id;
         }
 
-        @Specialization(guards = "isForeignObject(object)")
-        protected long objectIDForeign(Object object) {
-            return Integer.toUnsignedLong(hashCode(object));
-        }
-
-        @TruffleBoundary
-        private int hashCode(Object object) {
-            return object.hashCode();
+        @Specialization(guards = "isForeignObject(value)", limit = "getInteropCacheLimit()")
+        protected int objectIDForeign(Object value,
+                @CachedLibrary("value") InteropLibrary interop,
+                @Cached TranslateInteropExceptionNode translateInteropException) {
+            if (interop.hasIdentity(value)) {
+                try {
+                    return interop.identityHashCode(value);
+                } catch (UnsupportedMessageException e) {
+                    throw translateInteropException.execute(e);
+                }
+            } else {
+                return System.identityHashCode(value);
+            }
         }
 
         protected int getCacheLimit() {