+++ src/main/java/org/truffleruby/language/methods/LookupMethodNode.java
@@ -76,10 +76,6 @@ public abstract class LookupMethodNode extends RubyBaseNode {
 
         // Actual lookup
 
-        if (foreignProfile.profile(metaClass == coreLibrary().truffleInteropForeignClass)) {
-            return null;
-        }
-
         final DeclarationContext declarationContext = getDeclarationContext(frame, config);
         final InternalMethod method;
         // Lookup first in the metaclass as we are likely to find the method there
@@ -135,10 +131,6 @@ public abstract class LookupMethodNode extends RubyBaseNode {
     protected static MethodLookupResult lookupCached(RubyContext context, Frame callingFrame,
             RubyClass metaClass, String name, DispatchConfiguration config) {
         CompilerAsserts.neverPartOfCompilation("slow-path method lookup should not be compiled");
-
-        if (metaClass == context.getCoreLibrary().truffleInteropForeignClass) {
-            return new MethodLookupResult(null);
-        }
 
         final DeclarationContext declarationContext = getDeclarationContext(callingFrame, config);
         final MethodLookupResult method = ModuleOperations.lookupMethodCached(metaClass, name, declarationContext);